+++ src/connect/index.ts
@@ -42,4 +42,6 @@ export { ConnectionParams } from './ConnectionParams';
 export { IDiscovery } from './IDiscovery';
 export { ConnectionResolver } from './ConnectionResolver';
 export { MemoryDiscovery } from './MemoryDiscovery';
+export { CompositeConnectionResolver } from './CompositeConnectionResolver';
 export { DefaultDiscoveryFactory } from './DefaultDiscoveryFactory';
+export { ConnectionUtils } from './ConnectionUtils';