+++ CHANGELOG.md
@@ -1,5 +1,11 @@
 # <img src="https://uploads-ssl.webflow.com/5ea5d3315186cf5ec60c3ee4/5edf1c94ce4c859f2b188094_logo.svg" alt="Pip.Services Logo" width="200"> <br/> Component definitions for Node.js Changelog
 
+## <a name="3.1.0"></a> 3.1.0 (2021-03-26) 
+
+### Features
+* **connect** Added CompositeConnectionResolver class
+* **connect** Added ConnectionUtils class
+
 ## <a name="3.0.2"></a> 3.0.2 (2018-08-02) 
 
 ### Breaking changes